Implement a solution for issue #6 so that we can now convert to and from Foundation-based attribute dictionaries and arrays of Attributes. Separate Attribute.Name struct into its own file.
